Step 5: Move page allocator from kernel to sys
Move MetadataPageAllocator, PageAllocator trait, and all unsafe operations
to sys/src/memory/page_allocator.rs. Kernel re-exports from sys.
Remove dead align_down_ptr from kernel (now only used via sys).
