+++ sys/src/memory/mod.rs
@@ -1,5 +1,6 @@
 pub mod address;
 pub mod page;
+pub mod page_allocator;
 
 pub use address::{PhysAddr, VirtAddr};
 pub use page::{PAGE_SIZE, Page, Pages, PagesAsSlice, PinnedHeapPages};
